persist: pull Snapshot and impls into a new module
There's enough complexity here (and a bit more coming soon) that it's
warranted.
