+++ src/dataflow/src/source/mod.rs
@@ -13,7 +13,7 @@ use mz_avro::types::Value;
 use mz_dataflow_types::sources::AwsExternalId;
 use mz_dataflow_types::{DecodeError, SourceErrorDetails};
 use mz_persist::client::{StreamReadHandle, StreamWriteHandle};
-use mz_persist::indexed::Snapshot;
+use mz_persist::indexed::snapshot::Snapshot;
 use mz_persist::operators::stream::Persist;
 use mz_persist_types::Codec;
 use mz_repr::MessagePayload;
