+++ src/persist/src/golden_test.rs
@@ -18,7 +18,7 @@ use tracing::{debug, error, info};
 
 use crate::client::RuntimeClient;
 use crate::error::{Error, ErrorLog};
-use crate::indexed::Snapshot;
+use crate::indexed::snapshot::Snapshot;
 use crate::mem::{MemBlob, MemRegistry};
 use crate::nemesis::direct::{Direct, StartRuntime};
 use crate::nemesis::generator::{Generator, GeneratorConfig};
