+++ src/persist/src/operators/replay.rs
@@ -18,7 +18,7 @@ use timely::{Data as TimelyData, PartialOrder};
 
 use crate::client::DecodedSnapshot;
 use crate::error::Error;
-use crate::indexed::Snapshot;
+use crate::indexed::snapshot::Snapshot;
 use crate::operators::DEFAULT_OUTPUTS_PER_YIELD;
 
 /// Extension trait for [`Stream`].
